+++ Documentation/BuildInstructionsLadybird.md
@@ -7,7 +7,7 @@ Qt6 development packages, nasm, additional build tools, and a C++23 capable comp
 A Rust toolchain is also required. You can install it via [rustup](https://rustup.rs/).
 
 We currently use gcc-14 and clang-21 in our CI pipeline. If these versions are not available on your system, see
-[`Meta/find_compiler.py`](../Meta/find_compiler.py) for the minimum compatible version.
+[`find_compiler.py`](../Meta/Utils/find_compiler.py) for the minimum compatible version.
 
 CMake 3.30 or newer must be available in $PATH.
