+++ Meta/Utils/utils.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2025, Tim Flynn <trflynn89@ladybird.org>
+# Copyright (c) 2025-2026, Tim Flynn <trflynn89@ladybird.org>
 #
 # SPDX-License-Identifier: BSD-2-Clause
 
@@ -41,3 +41,19 @@ def run_command(
         return output.strip()
 
     return None
+
+
+def string_hash(string: str) -> int:
+    """Port of AK::string_hash that produces the same u32 value."""
+    h = 0
+
+    for ch in string:
+        h = (h + ord(ch)) & 0xFFFFFFFF
+        h = (h + (h << 10)) & 0xFFFFFFFF
+        h ^= h >> 6
+
+    h = (h + (h << 3)) & 0xFFFFFFFF
+    h ^= h >> 11
+    h = (h + (h << 15)) & 0xFFFFFFFF
+
+    return h
